fastapi-notes/19-bazaarapi/requirements.txt
# ============================================================
# BazaarAPI — Python Dependencies
# ============================================================
# Install with: pip install -r requirements.txt
# ============================================================

# Core framework
fastapi[standard]>=0.110.0
uvicorn[standard]>=0.27.0

# Database ORM
sqlmodel>=0.0.16

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# Configuration
pydantic-settings>=2.1.0

# HTTP client (for testing and external calls)
httpx>=0.27.0

# Payments (optional — mock client used if not installed)
razorpay>=1.4.1

# Migrations
alembic>=1.13.0
